pyproject.toml
[project]
name = "pkos"
version = "0.1.0"
description = ""
authors = [
    {name = "Siddharth Bariya",email = "sba.workspace@gmail.com"}
]
readme = "README.md"
package-mode = false
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12,<0.116.0",
    "uvicorn>=0.34.2,<0.35.0",
    "prisma @ git+https://github.com/RobertCraigie/prisma-client-py.git@v0.11.0",
    "asyncpg>=0.30.0,<0.31.0",
    "redis>=5.2.1,<6.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "pydantic==1.10.13",
    "prisma-client (>=0.2.1,<0.3.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "google-generativeai (>=0.8.5,<0.9.0)"
]




[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
pytest = "^8.3.5"
httpx = "^0.28.1"

